vix: extract shared value/support types out of oracle module
