+++ vix/src/oracle.rs
@@ -32,310 +32,13 @@ use crate::ast::Spanned;
 use crate::ast::{self, Arg, ArrayElem, Block, CommandPart, Expr, Member, PathRef, Pattern, Stmt};
 use crate::fetch::{FetchBackend, NoFetchBackend};
 use crate::module::{EnumInfo, ModuleTables, StructInfo, VariantShape, load_module_tables};
+use crate::value::deep_force_with;
 
-// ---------------------------------------------------------------------------
-// Values: hashable, totally ordered, canonical — and SHIPPABLE. Facet is the
-// data plane: a value serializes to postcard bytes and reconstitutes on
-// another host (or another oracle, which is the same thing minus the wire).
-// ---------------------------------------------------------------------------
-
-#[derive(facet::Facet, Debug, Clone)]
-#[repr(u8)]
-pub enum Value {
-    Int(i64),
-    Float(f64),
-    Bool(bool),
-    Str(String),
-    Path(String),
-    Flag(String),
-    Tuple(Vec<Value>),
-    Array(Vec<Value>),
-    Map(BTreeMap<Value, Value>),
-    /// Fields in DECLARATION order (that order is the total order).
-    Struct {
-        name: String,
-        fields: Vec<(String, Value)>,
-    },
-    /// `index` is the declaration index — the total order over variants.
-    Variant {
-        enum_name: String,
-        index: usize,
-        name: String,
-        payload: Payload,
-    },
-    /// A top-level function as a value.
-    Fn {
-        name: String,
-        hash: u64,
-    },
-    /// A closure: canonical body hash + captured environment.
-    Closure {
-        hash: u64,
-        params: Vec<String>,
-        body: Box<Expr>,
-        env: Vec<(String, Value)>,
-    },
-    /// A partially-applied named function (`f(a: x, ..)`).
-    Partial {
-        func: String,
-        given: Vec<(String, Value)>,
-    },
-    /// A content-addressed tree — the currency of exec composition. Reading a
-    /// Tree VALUE is pure (it's a value, not the world); only mounted trees
-    /// at exec time are observed.
-    Tree(crate::exec::Tree),
-    /// A tree STILL BEING PRODUCED by a live run. Projection (`pending / path`)
-    /// demands one path; identity (hash/cmp/ship) and whole-tree uses (mount,
-    /// glob, merge) force the flush. The id points into the run registry.
-    PendingTree {
-        run: u64,
-    },
-}
-
-#[derive(facet::Facet, Debug, Clone)]
-#[repr(u8)]
-pub enum Payload {
-    Unit,
-    Tuple(Vec<Value>),
-    Record(Vec<(String, Value)>),
-}
-
-impl Value {
-    fn rank(&self) -> u8 {
-        match self {
-            Value::Int(_) => 0,
-            Value::Float(_) => 1,
-            Value::Bool(_) => 2,
-            Value::Str(_) => 3,
-            Value::Path(_) => 4,
-            Value::Flag(_) => 5,
-            Value::Tuple(_) => 6,
-            Value::Array(_) => 7,
-            Value::Map(_) => 8,
-            Value::Struct { .. } => 9,
-            Value::Variant { .. } => 10,
-            Value::Fn { .. } => 11,
-            Value::Closure { .. } => 12,
-            Value::Partial { .. } => 13,
-            // Pending trees rank AS trees: identity forces them into one.
-            Value::Tree(_) | Value::PendingTree { .. } => 14,
-        }
-    }
-
-    /// Identity forces: a pending tree resolves to its flushed tree before
-    /// hashing/comparison. Panics if the registry lost the run (oracle-grade).
-    fn forced_tree(&self) -> Option<crate::exec::Tree> {
-        match self {
-            Value::Tree(t) => Some(t.clone()),
-            Value::PendingTree { run } => Some(force_run(*run).expect("pending run flushes").0),
-            _ => None,
-        }
-    }
-
-    /// A SHORT human rendering — event payloads say what is being computed
-    /// without dumping whole trees. Paths/strings verbatim, aggregates
-    /// summarized, trees by identity.
-    pub fn short(&self) -> String {
-        match self {
-            Value::Int(v) => v.to_string(),
-            Value::Float(v) => v.to_string(),
-            Value::Bool(v) => v.to_string(),
-            Value::Str(v) => format!("{v:?}"),
-            Value::Path(v) => v.clone(),
-            Value::Flag(v) => v.clone(),
-            Value::Tuple(vs) => format!(
-                "({})",
-                vs.iter().map(|v| v.short()).collect::<Vec<_>>().join(", ")
-            ),
-            Value::Array(vs) => format!(
-                "[{}]",
-                vs.iter().map(|v| v.short()).collect::<Vec<_>>().join(", ")
-            ),
-            Value::Map(entries) => format!("{{…{} entries}}", entries.len()),
-            Value::Struct { name, fields } => format!("{name}{{…{}}}", fields.len()),
-            Value::Variant {
-                enum_name, name, ..
-            } => format!("{enum_name}::{name}"),
-            Value::Fn { name, .. } => format!("fn {name}"),
-            Value::Closure { .. } => "closure".to_string(),
-            Value::Partial { func, .. } => format!("partial {func}"),
-            Value::Tree(t) => {
-                let mut h = DefaultHasher::new();
-                self.hash_into(&mut h);
-                format!("tree({:08x}, {} paths)", h.finish() as u32, t.entries.len())
-            }
-            Value::PendingTree { run } => format!("pending(run {run})"),
-        }
-    }
-
-    /// Structural hash — the canonical identity of a value.
-    pub fn hash_into(&self, h: &mut DefaultHasher) {
-        self.rank().hash(h);
-        match self {
-            Value::Int(v) => v.hash(h),
-            Value::Float(v) => normalize_float(*v).to_bits().hash(h),
-            Value::Bool(v) => v.hash(h),
-            Value::Str(v) | Value::Path(v) | Value::Flag(v) => v.hash(h),
-            Value::Tuple(vs) | Value::Array(vs) => {
-                vs.len().hash(h);
-                for v in vs {
-                    v.hash_into(h);
-                }
-            }
-            Value::Map(m) => {
-                m.len().hash(h);
-                for (k, v) in m {
-                    k.hash_into(h);
-                    v.hash_into(h);
-                }
-            }
-            Value::Struct { name, fields } => {
-                name.hash(h);
-                for (fname, v) in fields {
-                    fname.hash(h);
-                    v.hash_into(h);
-                }
-            }
-            Value::Variant {
-                enum_name,
-                index,
-                payload,
-                ..
-            } => {
-                enum_name.hash(h);
-                index.hash(h);
-                match payload {
-                    Payload::Unit => {}
-                    Payload::Tuple(vs) => {
-                        for v in vs {
-                            v.hash_into(h);
-                        }
-                    }
-                    Payload::Record(fs) => {
-                        for (n, v) in fs {
-                            n.hash(h);
-                            v.hash_into(h);
-                        }
-                    }
-                }
-            }
-            Value::Fn { hash, .. } => hash.hash(h),
-            Value::Closure { hash, env, .. } => {
-                hash.hash(h);
-                for (n, v) in env {
-                    n.hash(h);
-                    v.hash_into(h);
-                }
-            }
-            Value::Partial { func, given } => {
-                func.hash(h);
-                for (n, v) in given {
-                    n.hash(h);
-                    v.hash_into(h);
-                }
-            }
-            Value::Tree(_) | Value::PendingTree { .. } => {
-                self.forced_tree().expect("tree rank").fingerprint().hash(h)
-            }
-        }
-    }
-
-    pub fn canon_hash(&self) -> u64 {
-        let mut h = DefaultHasher::new();
-        self.hash_into(&mut h);
-        h.finish()
-    }
-}
-
-/// Total order on floats: -0.0 == 0.0, NaN normalized and sorted last.
-fn normalize_float(v: f64) -> f64 {
-    if v.is_nan() {
-        f64::NAN // one canonical NaN bit pattern via the constant
-    } else if v == 0.0 {
-        0.0
-    } else {
-        v
-    }
-}
-
-fn float_cmp(a: f64, b: f64) -> std::cmp::Ordering {
-    normalize_float(a).total_cmp(&normalize_float(b))
-}
-
-impl PartialEq for Value {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other) == std::cmp::Ordering::Equal
-    }
-}
-impl Eq for Value {}
-
-impl PartialOrd for Value {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for Value {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        use std::cmp::Ordering;
-        match (self, other) {
-            (Value::Int(a), Value::Int(b)) => a.cmp(b),
-            (Value::Float(a), Value::Float(b)) => float_cmp(*a, *b),
-            (Value::Bool(a), Value::Bool(b)) => a.cmp(b),
-            (Value::Str(a), Value::Str(b))
-            | (Value::Path(a), Value::Path(b))
-            | (Value::Flag(a), Value::Flag(b)) => a.cmp(b),
-            (Value::Tuple(a), Value::Tuple(b)) | (Value::Array(a), Value::Array(b)) => a.cmp(b),
-            (Value::Map(a), Value::Map(b)) => a.cmp(b),
-            (
-                Value::Struct {
-                    name: an,
-                    fields: af,
-                },
-                Value::Struct {
-                    name: bn,
-                    fields: bf,
-                },
-            ) => an.cmp(bn).then_with(|| af.cmp(bf)),
-            (
-                Value::Variant {
-                    enum_name: ae,
-                    index: ai,
-                    payload: ap,
-                    ..
-                },
-                Value::Variant {
-                    enum_name: be,
-                    index: bi,
-                    payload: bp,
-                    ..
-                },
-            ) => ae.cmp(be).then_with(|| ai.cmp(bi)).then_with(|| {
-                // Same variant: compare payloads structurally.
-                match (ap, bp) {
-                    (Payload::Unit, Payload::Unit) => Ordering::Equal,
-                    (Payload::Tuple(a), Payload::Tuple(b)) => a.cmp(b),
-                    (Payload::Record(a), Payload::Record(b)) => a.cmp(b),
-                    _ => Ordering::Equal, // same variant index implies same shape
-                }
-            }),
-            (Value::Fn { hash: a, .. }, Value::Fn { hash: b, .. }) => a.cmp(b),
-            (Value::Closure { .. }, Value::Closure { .. })
-            | (Value::Partial { .. }, Value::Partial { .. }) => {
-                self.canon_hash().cmp(&other.canon_hash())
-            }
-            (
-                Value::Tree(_) | Value::PendingTree { .. },
-                Value::Tree(_) | Value::PendingTree { .. },
-            ) => {
-                let a = self.forced_tree().expect("tree rank");
-                let b = other.forced_tree().expect("tree rank");
-                a.entries.cmp(&b.entries)
-            }
-            _ => self.rank().cmp(&other.rank()),
-        }
-    }
-}
+pub use crate::support::{ExecBackend, PathDemand, PathMissing, PathPending, PendingRun};
+pub(crate) use crate::support::{
+    LocalRun, assign_roles, force_run, runs, splice_into, subtree, tool_for,
+};
+pub use crate::value::{Payload, Value, deep_force, receive, ship};
 
 // ---------------------------------------------------------------------------
 // Canonical AST identity: the generated `strip_spans` zeroes every span
@@ -350,82 +53,6 @@ fn canon_expr_hash(expr: &Expr) -> u64 {
     let mut h = DefaultHasher::new();
     bytes.hash(&mut h);
     h.finish()
-}
-
-/// Serialize a value for transport — the exec primitive's payload format.
-/// Shipping is an IDENTITY use: pending trees force first.
-pub fn ship(value: &Value) -> Result<Vec<u8>, String> {
-    let forced = deep_force(value.clone())?;
-    phon::api::encode(&forced).map_err(|e| format!("ship: {e}"))
-}
-
-/// Replace every pending tree in a value with its flushed tree (identity
-/// forces, recursively). The edge of the graph — `Oracle::call` — does this
-/// to its result: DEMAND ENTERS AT THE EDGE.
-pub fn deep_force(value: Value) -> Result<Value, String> {
-    deep_force_with(value, &|id| force_run(id).map(|(t, _)| t))
-}
-
-fn deep_force_with(
-    value: Value,
-    force: &impl Fn(u64) -> Result<crate::exec::Tree, String>,
-) -> Result<Value, String> {
-    Ok(match value {
-        Value::PendingTree { run } => Value::Tree(force(run)?),
-        Value::Tuple(vs) => Value::Tuple(
-            vs.into_iter()
-                .map(|v| deep_force_with(v, force))
-                .collect::<Result<_, _>>()?,
-        ),
-        Value::Array(vs) => Value::Array(
-            vs.into_iter()
-                .map(|v| deep_force_with(v, force))
-                .collect::<Result<_, _>>()?,
-        ),
-        Value::Map(m) => Value::Map(
-            m.into_iter()
-                .map(|(k, v)| {
-                    Ok::<_, String>((deep_force_with(k, force)?, deep_force_with(v, force)?))
-                })
-                .collect::<Result<_, _>>()?,
-        ),
-        Value::Struct { name, fields } => Value::Struct {
-            name,
-            fields: fields
-                .into_iter()
-                .map(|(n, v)| Ok::<_, String>((n, deep_force_with(v, force)?)))
-                .collect::<Result<_, _>>()?,
-        },
-        Value::Variant {
-            enum_name,
-            index,
-            name,
-            payload,
-        } => Value::Variant {
-            enum_name,
-            index,
-            name,
-            payload: match payload {
-                Payload::Unit => Payload::Unit,
-                Payload::Tuple(vs) => Payload::Tuple(
-                    vs.into_iter()
-                        .map(|v| deep_force_with(v, force))
-                        .collect::<Result<_, _>>()?,
-                ),
-                Payload::Record(fs) => Payload::Record(
-                    fs.into_iter()
-                        .map(|(n, v)| Ok::<_, String>((n, deep_force_with(v, force)?)))
-                        .collect::<Result<_, _>>()?,
-                ),
-            },
-        },
-        other => other,
-    })
-}
-
-/// Reconstitute a shipped value on the receiving side.
-pub fn receive(bytes: &[u8]) -> Result<Value, String> {
-    phon::api::decode(bytes).map_err(|e| format!("receive: {e}"))
 }
 
 // ---------------------------------------------------------------------------
@@ -498,141 +125,6 @@ pub enum Event {
 pub type EventSink = Box<dyn Fn(u64, &Event) + Send>;
 
 type Frame = Vec<(String, Value)>;
-
-/// Where command blocks actually run. The oracle's local two-tier cache is
-/// the default; a fleet of wire executors plugs in here (sync on purpose —
-/// the oracle is sync and wasm-clean; the async wire bridges on its side).
-///
-/// DEMAND-DRIVEN: spawn returns immediately with a live run. The language
-/// rule is "projection doesn't force; identity does" — demanding one path
-/// blocks only until the producer writes it; hashing/comparing/shipping/
-/// mounting a pending tree forces its flush.
-pub trait ExecBackend: Send + Sync {
-    fn spawn(
-        &self,
-        command: &str,
-        plan: &crate::exec::ExecPlan,
-        capability: u64,
-        mounts: &[crate::exec::Mount],
-    ) -> Result<std::sync::Arc<dyn PendingRun>, String>;
-}
-
-/// Result of demanding one path from a tree-producing run.
-///
-/// `PendingRun::demand_path` blocks while the producer is still capable of
-/// writing the requested file. Once it returns, the path is either available
-/// as a single file, known absent from the finished tree, or requires the
-/// finished tree to decide a directory projection.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum PathDemand {
-    File(String),
-    FinishRequired(PathPending),
-    Missing(PathMissing),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PathPending {
-    pub path: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PathMissing {
-    pub path: String,
-}
-
-impl PathPending {
-    fn new(path: &str) -> Self {
-        Self {
-            path: path.to_string(),
-        }
-    }
-}
-
-impl PathMissing {
-    fn new(path: &str) -> Self {
-        Self {
-            path: path.to_string(),
-        }
-    }
-
-    pub fn diagnostic(&self) -> String {
-        format!("path `{}` is missing from resolved tree", self.path)
-    }
-}
-
-/// A live (or completed) run. Values hold these through the run REGISTRY so
-/// Value stays plain facet data (a run id), and context-free forcing (Ord,
-/// Hash, ship) still works.
-pub trait PendingRun: Send + Sync {
-    /// Block until the producer writes this path (or fails/finishes without it).
-    fn demand_path(&self, path: &str) -> Result<PathDemand, String>;
-    /// Block until completion; idempotent. Returns the flushed tree and how
-    /// the run was served.
-    fn flush(&self) -> Result<(crate::exec::Tree, crate::exec::ExecEvent), String>;
-}
-
-/// The process-global run registry: PendingTree values carry a u64 into this
-/// table. Oracle-grade shortcut (the real engine owns runs in its graph);
-/// std-only so the crate stays wasm-clean — blocking lives in backends.
-pub(crate) mod runs {
-    use std::collections::HashMap;
-    use std::sync::{Arc, Mutex, OnceLock};
-
-    use super::PendingRun;
-
-    fn table() -> &'static Mutex<HashMap<u64, Arc<dyn PendingRun>>> {
-        static TABLE: OnceLock<Mutex<HashMap<u64, Arc<dyn PendingRun>>>> = OnceLock::new();
-        TABLE.get_or_init(|| Mutex::new(HashMap::new()))
-    }
-
-    pub fn register(run: Arc<dyn PendingRun>) -> u64 {
-        use std::sync::atomic::{AtomicU64, Ordering};
-        static NEXT: AtomicU64 = AtomicU64::new(1);
-        let id = NEXT.fetch_add(1, Ordering::SeqCst);
-        table().lock().unwrap().insert(id, run);
-        id
-    }
-
-    pub fn get(id: u64) -> Option<Arc<dyn PendingRun>> {
-        table().lock().unwrap().get(&id).cloned()
-    }
-}
-
-/// Force a pending tree to its flushed value ("identity forces").
-fn force_run(id: u64) -> Result<(crate::exec::Tree, crate::exec::ExecEvent), String> {
-    runs::get(id)
-        .ok_or_else(|| format!("pending run {id} not in the registry"))?
-        .flush()
-}
-
-/// The local cache's "pending" run: already complete at spawn time (the
-/// local path is the caching oracle; the wire is where demand overlaps).
-pub(crate) struct LocalRun {
-    pub(crate) outputs: crate::exec::Tree,
-    pub(crate) event: crate::exec::ExecEvent,
-}
-
-impl PendingRun for LocalRun {
-    fn demand_path(&self, path: &str) -> Result<PathDemand, String> {
-        if let Some(contents) = self.outputs.entries.get(path) {
-            return Ok(PathDemand::File(contents.clone()));
-        }
-        let prefix = format!("{path}/");
-        if self
-            .outputs
-            .entries
-            .keys()
-            .any(|entry| entry.starts_with(&prefix))
-        {
-            return Ok(PathDemand::FinishRequired(PathPending::new(path)));
-        }
-        Ok(PathDemand::Missing(PathMissing::new(path)))
-    }
-
-    fn flush(&self) -> Result<(crate::exec::Tree, crate::exec::ExecEvent), String> {
-        Ok((self.outputs.clone(), self.event.clone()))
-    }
-}
 
 pub struct Oracle {
     fns: HashMap<String, ast::FnItem>,
@@ -1815,122 +1307,6 @@ fn path_names_variant(path: &PathRef, enum_name: &str, variant: &str) -> bool {
 
 fn path_names_struct(path: &PathRef, name: &str) -> bool {
     matches!(path, PathRef::Identifier(n) if n.value == name)
-}
-
-/// Tree / Path: a directory re-rooted, or one file cut out by basename.
-pub(crate) fn subtree(tree: &crate::exec::Tree, path: &str) -> Result<crate::exec::Tree, String> {
-    if let Some(contents) = tree.entries.get(path) {
-        let base = path.rsplit_once('/').map(|(_, b)| b).unwrap_or(path);
-        return Ok(crate::exec::Tree::of(&[(base, contents.as_str())]));
-    }
-    let prefix = format!("{path}/");
-    let entries: BTreeMap<String, String> = tree
-        .entries
-        .iter()
-        .filter_map(|(k, v)| k.strip_prefix(&prefix).map(|r| (r.to_string(), v.clone())))
-        .collect();
-    if entries.is_empty() {
-        return Err(format!("no `{path}` in tree"));
-    }
-    Ok(crate::exec::Tree { entries })
-}
-
-/// Splice a value into a command: paths/strings/flags become argv text,
-/// arrays flatten, and TREES become mounts. Mount paths are ROLE-STABLE
-/// (/m/0, /m/1 by splice order): the plan is the identity, the world is what
-/// varies — a content-derived path would leak the world into the identity
-/// and tier-2 could never match. A single-file tree splices as the file's
-/// path inside its mount; a directory splices as the mount root.
-pub(crate) fn splice_into(
-    value: Value,
-    argv: &mut Vec<String>,
-    mounts: &mut Vec<crate::exec::Mount>,
-) -> Result<(), String> {
-    match value {
-        Value::Path(p) | Value::Str(p) | Value::Flag(p) => argv.push(p),
-        Value::Int(i) => argv.push(i.to_string()),
-        Value::Float(f) => argv.push(f.to_string()),
-        Value::Array(vs) => {
-            for v in vs {
-                splice_into(v, argv, mounts)?;
-            }
-        }
-        Value::Tree(t) => {
-            let root = format!("/m/{}", mounts.len());
-            let text = if t.entries.len() == 1 {
-                format!("{root}/{}", t.entries.keys().next().unwrap())
-            } else {
-                root.clone()
-            };
-            mounts.push(crate::exec::Mount { at: root, tree: t });
-            argv.push(text);
-        }
-        other => return Err(format!("cannot splice {other:?} into a command")),
-    }
-    Ok(())
-}
-
-/// The toy role tables — stand-ins for snark command grammars (which will
-/// assign these roles from real grammar productions, versioned with the
-/// capability).
-pub(crate) fn assign_roles(
-    command: &str,
-    argv: &[String],
-) -> Result<crate::exec::ExecPlan, String> {
-    use crate::exec::Role;
-    let mut out = Vec::new();
-    match command {
-        "cc" => {
-            let mut prev: Option<&str> = None;
-            for arg in argv {
-                let role = match prev {
-                    Some("-o") => Role::Output,
-                    Some("-I") => Role::SearchDir,
-                    _ if arg.starts_with("/m/") => Role::Input,
-                    _ => Role::Flag,
-                };
-                out.push((arg.clone(), role));
-                prev = Some(arg.as_str());
-            }
-        }
-        "ar" => {
-            for arg in argv {
-                let role = if arg.starts_with("/m/") {
-                    crate::exec::Role::Input
-                } else if arg == "rcs" {
-                    Role::Flag
-                } else {
-                    Role::Output
-                };
-                out.push((arg.clone(), role));
-            }
-        }
-        "rustc" => {
-            let mut prev: Option<&str> = None;
-            for arg in argv {
-                let role = match prev {
-                    Some("-o") => Role::Output,
-                    Some("-L") => Role::SearchDir,
-                    _ if arg.starts_with("/m/") => Role::Input,
-                    _ if arg.starts_with("-L") && arg.contains("/m/") => Role::SearchDir,
-                    _ => Role::Flag,
-                };
-                out.push((arg.clone(), role));
-                prev = Some(arg.as_str());
-            }
-        }
-        other => return Err(format!("no command grammar for `{other}`")),
-    }
-    Ok(crate::exec::ExecPlan { argv: out })
-}
-
-pub(crate) fn tool_for(command: &str) -> Result<&'static dyn crate::exec::Tool, String> {
-    match command {
-        "cc" => Ok(&crate::exec::FakeCc),
-        "ar" => Ok(&crate::exec::FakeAr),
-        "rustc" => Ok(&crate::exec::FakeRustc),
-        other => Err(format!("no tool for `{other}`")),
-    }
 }
 
 fn parse_number(text: &str) -> Value {
